+++ src/themes/greenyBlueTheme/index.js
@@ -1,0 +1,17 @@
+const greenyBlueTheme = {
+  colors: {
+    lightPrimary: '#00b4f1',
+    primary: '#00a5e1',
+    darkPrimary: '#0099d1',
+    lightSecondary: '#a5d604',
+    secondary: '#9ecd04',
+    darkSecondary: '#96c304',
+    black: '#333',
+    smoke: '#5b5b5b',
+    darkSmoke: '#8a8a8a',
+    snow: '#F8F8F8',
+    darkSnow: '#dfdfdf',
+  },
+}
+
+export default greenyBlueTheme